+++ x/ecocredit/client/testsuite/query.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 
 	"github.com/cosmos/cosmos-sdk/client/flags"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/regen-network/regen-ledger/types/testutil/cli"
 	coreclient "github.com/regen-network/regen-ledger/x/ecocredit/client"
@@ -565,23 +566,25 @@ func (s *IntegrationTestSuite) TestQueryCreditTypesCmd() {
 	}
 }
 
-// TODO: #1363
-// func (s *IntegrationTestSuite) TestQueryParamsCmd() {
-// 	val := s.network.Validators[0]
-// 	clientCtx := val.ClientCtx
-// 	clientCtx.OutputFormat = "JSON"
-// 	require := s.Require()
+func (s *IntegrationTestSuite) TestQueryParamsCmd() {
+	val := s.network.Validators[0]
+	clientCtx := val.ClientCtx
+	clientCtx.OutputFormat = "JSON"
+	require := s.Require()
 
-// 	cmd := coreclient.QueryParamsCmd()
-// 	out, err := cli.ExecTestCLICmd(clientCtx, cmd, []string{})
-// 	require.NoError(err)
+	cmd := coreclient.QueryParamsCmd()
+	out, err := cli.ExecTestCLICmd(clientCtx, cmd, []string{})
+	require.NoError(err)
 
-// 	var params core.QueryParamsResponse
-// 	s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), &params))
-// 	require.NoError(err)
+	var params core.QueryParamsResponse
+	s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), &params))
+	require.NoError(err)
 
-// 	require.Equal(core.DefaultParams(), *params.Params)
-// }
+	require.Equal(core.DefaultParams().BasketFee, params.Params.BasketFee)
+	require.Equal(core.DefaultParams().CreditClassFee, params.Params.CreditClassFee)
+	require.False(params.Params.AllowlistEnabled)
+	require.Equal([]string{sdk.AccAddress("issuer1").String(), sdk.AccAddress("issuer2").String()}, params.Params.AllowedClassCreators)
+}
 
 func (s *IntegrationTestSuite) TestQueryProjectsCmd() {
 	require := s.Require()
@@ -865,4 +868,52 @@ func (s *IntegrationTestSuite) TestQueryCreditTypeCmd() {
 			}
 		})
 	}
+}
+
+func (s *IntegrationTestSuite) TestQueryAllowedClassCreators() {
+	val := s.network.Validators[0]
+	clientCtx := val.ClientCtx
+	clientCtx.OutputFormat = outputFormat
+
+	cmd := coreclient.QueryAllowedClassCreatorsCmd()
+	out, err := cli.ExecTestCLICmd(clientCtx, cmd, []string{})
+
+	s.Require().NoError(err, out.String())
+
+	var res core.QueryAllowedClassCreatorsResponse
+	s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), &res))
+	s.Require().Len(res.ClassCreators, 2)
+	s.Require().Equal(res.ClassCreators[0], sdk.AccAddress("issuer1").String())
+	s.Require().Equal(res.ClassCreators[1], sdk.AccAddress("issuer2").String())
+}
+
+func (s *IntegrationTestSuite) TestQueryCreditClassAllowlistEnableCmd() {
+	val := s.network.Validators[0]
+	clientCtx := val.ClientCtx
+	clientCtx.OutputFormat = outputFormat
+
+	cmd := coreclient.QueryCreditClassAllowlistEnabledCmd()
+	out, err := cli.ExecTestCLICmd(clientCtx, cmd, []string{})
+
+	s.Require().NoError(err, out.String())
+
+	var res core.QueryCreditClassAllowlistEnabledResponse
+	s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), &res))
+	s.Require().False(res.AllowlistEnabled)
+}
+
+func (s *IntegrationTestSuite) TestQueryCreditClassFeesCmd() {
+	val := s.network.Validators[0]
+	clientCtx := val.ClientCtx
+	clientCtx.OutputFormat = outputFormat
+
+	cmd := coreclient.QueryCreditClassFeesCmd()
+	out, err := cli.ExecTestCLICmd(clientCtx, cmd, []string{})
+
+	s.Require().NoError(err, out.String())
+
+	var res core.QueryCreditClassFeesResponse
+	s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), &res))
+	s.Require().Equal(res.Fees.Len(), 1)
+	s.Require().Equal(res.Fees.AmountOf(sdk.DefaultBondDenom), core.DefaultCreditClassFee)
 }