+++ x/ecocredit/server/server.go
@@ -68,7 +68,7 @@ func NewServer(storeKey storetypes.StoreKey, legacySubspace paramtypes.Subspace,
 	s.stateStore = coreStore
 	s.basketStore = basketStore
 	s.marketplaceStore = marketStore
-	s.CoreKeeper = core.NewKeeper(coreStore, bankKeeper, coreAddr, basketStore, authority)
+	s.CoreKeeper = core.NewKeeper(coreStore, bankKeeper, coreAddr, basketStore, marketStore, authority)
 	s.BasketKeeper = basketkeeper.NewKeeper(basketStore, coreStore, bankKeeper, s.legacySubspace, basketAddr, authority)
 	s.MarketplaceKeeper = marketkeeper.NewKeeper(marketStore, coreStore, bankKeeper, s.legacySubspace, authority)